requirements.txt
flask~=3.0.0
pyjwt~=2.8.0
python-dotenv~=1.0.0
pymongo
bcrypt~=4.0.1
tinydb~=4.8.0
torch~=2.1.0
opencv-python~=4.8.1.78
pandas~=2.1.1
matplotlib~=3.8.0
numpy~=1.26.1
Pillow~=10.1.0
torchvision~=0.16.0
tqdm~=4.66.1
pyyaml~=6.0.1
seaborn~=0.13.0
netifaces
requests~=2.31.0
jwt~=1.3.1
cryptography~=41.0.4